// src/main/typescript/src/models/HealthResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * XStr.me API
 * OpenAPI specification for XStr.me services.  This API provides endpoints for managing and accessing XStr.me functionality. 
 *
 * The version of the OpenAPI document: 0.0.1-alpha
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface HealthResponse
 */
export interface HealthResponse {
    /**
     * Health status of the API
     * @type {string}
     * @memberof HealthResponse
     */
    status: HealthResponseStatusEnum;
    /**
     * Timestamp of the health check
     * @type {Date}
     * @memberof HealthResponse
     */
    timestamp: Date;
    /**
     * API version
     * @type {string}
     * @memberof HealthResponse
     */
    version?: string;
    /**
     * API uptime duration
     * @type {string}
     * @memberof HealthResponse
     */
    uptime?: string;
}

/**
* @export
* @enum {string}
*/
export enum HealthResponseStatusEnum {
    HEALTHY = 'healthy',
    UNHEALTHY = 'unhealthy'
}


/**
 * Check if a given object implements the HealthResponse interface.
 */
export function instanceOfHealthResponse(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "status" in value;
    isInstance = isInstance && "timestamp" in value;

    return isInstance;
}

export function HealthResponseFromJSON(json: any): HealthResponse {
    return HealthResponseFromJSONTyped(json, false);
}

export function HealthResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): HealthResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'status': json['status'],
        'timestamp': (new Date(json['timestamp'])),
        'version': !exists(json, 'version') ? undefined : json['version'],
        'uptime': !exists(json, 'uptime') ? undefined : json['uptime'],
    };
}

export function HealthResponseToJSON(value?: HealthResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'status': value.status,
        'timestamp': (value.timestamp.toISOString()),
        'version': value.version,
        'uptime': value.uptime,
    };
}